+++ packages/drizzle-db/src/migrate.ts
@@ -1,25 +1,10 @@
-// import { neon } from "@neondatabase/serverless";
-// import { config } from "dotenv";
-// import { drizzle } from "drizzle-orm/neon-http"; //import the drizzle that work for neon serverless driver
-// import { migrate } from "drizzle-orm/neon-http/migrator";
+import "dotenv/config";
 
-// config({ path: "../../.env" });
+import { migrate } from "drizzle-orm/postgres-js/migrator";
 
-// const DATABASE_URL = process.env.DATABASE_URL ?? "";
+import { client, db } from "./index";
 
-// console.log(DATABASE_URL);
-// const sql = neon(DATABASE_URL); //estabilish connection to db
-// const db = drizzle(sql); //wrapping db connection with drizzle, this gave us actual query builder
-
-// const main = async () => {
-//   try {
-//     await migrate(db, {
-//       migrationsFolder: "migrations",
-//     });
-//     console.log("Migration successful");
-//   } catch (error) {
-//     console.log(error);
-//     process.exit(1);
-//   }
-// };
-// await main();
+// This will run migrations on the database, skipping the ones already applied
+await migrate(db, { migrationsFolder: "../migrations" });
+// Don't forget to close the connection, otherwise the script will hang
+await client.end();
