--- packages/drizzle-db/src/models/schema.ts
@@ -1,9 +1,0 @@
-export { dish, dishRelations } from "./dish";
-export { dietRestriction } from "./dietRestriction";
-export { event } from "./event";
-export { menu, menuRelations } from "./menu";
-export { menuPeriod, menuPeriodName, menuPeriodRelations } from "./menuPeriod";
-export { nutritionInfo } from "./nutritionInfo";
-export { restaurant, restaurantName, restaurantRelations } from "./restaurant";
-export { station, stationRelations } from "./station";
-export { pushToken } from "./pushToken";
