--- packages/drizzle-db/src/schema/users.ts
@@ -1,5 +1,0 @@
-import { pgTable, text } from "drizzle-orm/pg-core";
-
-export const pushToken = pgTable("PushToken", {
-  token: text("token").primaryKey().notNull(),
-});
